ensure a non-nil error message for invalid emails and phone numbers
I found that I needed to add this because my email validity tests were
failing without it. using shoulda-matchers, I wrote the following test
with rspec:

```
it { is_expected.not_to allow_value('foo').for(:contact_email) }
```

the test failed when shoulda-matchers raised a TypeError with message
`can't dup NilClass`. tracing this I found that after calling `valid?`
on my record, the AM::Errors object had this error:

```
:contact_email => [
 [0] nil
]
```

for some reason passing a nil message to AM:Errors#add was not resulting
in the default `:invalid` message. this patch, however, makes things
work as expected.

it's not clear to me why the invalid error message test in this library
passed before. but it's still passing now, and my application works as
expected.